+++ src/effects/auras/_module.mjs
@@ -1,3 +1,3 @@
 export { default as test } from "./test.mjs";
 export { default as ConditionOnEntry } from "./ConditionOnEntry.mjs";
-// export { default as DamageOnEntry } from "./DamageOnEntry.mjs";
+export { default as DamageOnEntry } from "./DamageOnEntry.mjs";