+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - Monsters parsed in 6.0.x might have concentration removed from spells added.
+- Midi/ActiveAuras spell automation fixes and improvements for Grease, and Create Bonfire.
 
 # 6.0.63