+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Up
+# 6.0.64
 
 - Monsters parsed in 6.0.x might have concentration removed from spells added.
 - Midi/ActiveAuras spell automation fixes and improvements for Grease, Incendiary Cloud, Cloudkill, and Create Bonfire.